requirements.txt
requests
Flask
colorthief
Pillow
tqdm